大幅重构AM API

See merge request !90

+++ am/arch/x86-qemu/src/mpe.cpp
@@ -11,39 +11,39 @@ extern "C" {
 #define MP_MAGIC 0x5f504d5f // _MP_
 
 struct MPConf {     // configuration table header
-  u8 signature[4];       // "PCMP"
-  u16 length;        // total table length
-  u8 version;        // [14]
-  u8 checksum;         // all bytes must add up to 0
-  u8 product[20];      // product id
-  u32 *oemtable;         // OEM table pointer
-  u16 oemlength;       // OEM table length
-  u16 entry;         // entry count
-  u32 *lapicaddr;        // address of local APIC
-  u16 xlength;         // extended table length
-  u8 xchecksum;        // extended table checksum
-  u8 reserved;
+  uint8_t signature[4];       // "PCMP"
+  uint16_t length;        // total table length
+  uint8_t version;        // [14]
+  uint8_t checksum;         // all bytes must add up to 0
+  uint8_t product[20];      // product id
+  uint32_t *oemtable;         // OEM table pointer
+  uint16_t oemlength;       // OEM table length
+  uint16_t entry;         // entry count
+  uint32_t *lapicaddr;        // address of local APIC
+  uint16_t xlength;         // extended table length
+  uint8_t xchecksum;        // extended table checksum
+  uint8_t reserved;
 };
 typedef struct MPConf MPConf;
 
 struct MPDesc {
   int magic;
   MPConf *conf; // MP config table addr
-  u8 length; // 1
-  u8 specrev; // [14]
-  u8 checksum; // all bytes add to 0
-  u8 type;   // config type
-  u8 imcrp;
-  u8 reserved[3];
+  uint8_t length; // 1
+  uint8_t specrev; // [14]
+  uint8_t checksum; // all bytes add to 0
+  uint8_t type;   // config type
+  uint8_t imcrp;
+  uint8_t reserved[3];
 };
 typedef struct MPDesc MPDesc;
 
 int _NR_CPU = 0;
-extern u32 *lapic;
+extern uint32_t *lapic;
 
 static MPDesc *search() {
   for (char *st = (char*)0xf0000; st != (char*)0xffffff; st ++) {
-    if (*(u32*)st == MP_MAGIC) { // starts with magic _MP_
+    if (*(uint32_t*)st == MP_MAGIC) { // starts with magic _MP_
       MPDesc *mp = (MPDesc*)st;
       char checksum = 0;
       for (char *p = st; p != st + sizeof(MPDesc); p ++) {
@@ -77,7 +77,7 @@ void smp_init() {
 
 static void (* volatile _entry)();
 
-static ulong ap_boot = 0;
+static intptr_t ap_boot = 0;
 
 static void mp_entry() {
   if (_cpu() != 0) {
@@ -92,11 +92,11 @@ void _mpe_init(void (*entry)()) {
   _entry = entry;
 
   for (int cpu = 1; cpu < _NR_CPU; cpu ++) {
-    *(u16*)(0x7c00 + 510) = 0x55aa;
-    *reinterpret_cast<u32*>(0x7000) = 0x007c00ea;  // code for ljmp
-    *reinterpret_cast<u32*>(0x7004) = 0x00000000;
+    *(uint16_t*)(0x7c00 + 510) = 0x55aa;
+    *reinterpret_cast<uint32_t*>(0x7000) = 0x007c00ea;  // code for ljmp
+    *reinterpret_cast<uint32_t*>(0x7004) = 0x00000000;
     *(void**)(0x7010) = (void*)mp_entry;
-    *reinterpret_cast<u32*>(0x7020) += 1024; // 1KB bootstrap stack
+    *reinterpret_cast<uint32_t*>(0x7020) += 1024; // 1KB bootstrap stack
     lapic_bootap(cpu, 0x7000);
     while (_atomic_xchg(&ap_boot, 0) != 1);
   }
@@ -104,8 +104,8 @@ void _mpe_init(void (*entry)()) {
   mp_entry();
 }
 
-ulong _atomic_xchg(volatile ulong *addr, ulong newval) {
-  ulong result;
+intptr_t _atomic_xchg(volatile intptr_t *addr, intptr_t newval) {
+  intptr_t result;
   asm volatile("lock; xchgl %0, %1":
     "+m"(*addr), "=a"(result): "1"(newval): "cc");
   return result;
